+++ src/main/java/com/github/gtexpert/gtwp/integration/traverse/recipes/TraverseWoodRecipe.java
@@ -1,0 +1,38 @@
+package com.github.gtexpert.gtwp.integration.traverse.recipes;
+
+import java.util.Arrays;
+import java.util.List;
+
+import gregtech.loaders.WoodTypeEntry;
+
+import com.github.gtexpert.gtwp.api.util.Mods;
+import com.github.gtexpert.gtwp.loaders.GTWPWoodRecipeLoader;
+
+public class TraverseWoodRecipe {
+
+    private static List<WoodTypeEntry> DEFAULT_ENTRIES;
+
+    private static List<WoodTypeEntry> getDefaultEntries() {
+        if (DEFAULT_ENTRIES == null) {
+            final String mcModId = Mods.Names.TRAVERES;
+            return DEFAULT_ENTRIES = Arrays.asList(
+                    new WoodTypeEntry.Builder(mcModId, "silverwood")
+                            .log(Mods.Traverse.getItem("fir_log", 1)).removeCharcoalRecipe()
+                            .planks(Mods.Traverse.getItem("fir_planks", 1), "fir_planks")
+                            .slab(Mods.Traverse.getItem("fir_slab", 1), "fir_slab")
+                            .stairs(Mods.Traverse.getItem("fir_stairs", 1), "fir_stairs")
+                            .build());
+        }
+        return DEFAULT_ENTRIES;
+    }
+
+    public static void init() {
+        for (WoodTypeEntry entry : getDefaultEntries()) {
+            GTWPWoodRecipeLoader.removePlankRecipe(false, entry);
+
+            GTWPWoodRecipeLoader.registerWoodTypeRecipe(false, entry);
+            GTWPWoodRecipeLoader.addCuttingRecipe(entry);
+            GTWPWoodRecipeLoader.addSawmillRecipe(entry);
+        }
+    }
+}